requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
open_clip_torch>=2.20.0
pillow>=9.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pandas>=2.0.0
umap-learn>=0.5.0
salesforce-lavis @ git+https://github.com/salesforce/LAVIS.git